requirements.txt
flask==2.2.3
python-telegram-bot[job-queue]==20.1
nest-asyncio==1.5.6
gevent==23.9.1
python-dotenv==0.21.1
werkzeug>=2.2.2,<3.0
decouple==0.0.7
gunicorn
waitress
aiohttp 

